fix the tests
