+++ tests/grpc/gov.go
@@ -209,29 +209,3 @@ func MakeVoteAndCheckProposal(umeeClient client.Client, resp sdk.TxResponse) err
 
 	return fmt.Errorf("proposal %d failed to pass with status: %s", proposalIDInt, propStatus)
 }
-
-func GetTxResponse(umeeClient client.Client, txHash string) (resp *sdk.TxResponse, err error) {
-	for i := 0; i < 5; i++ {
-		resp, err = umeeClient.QueryTxHash(txHash)
-		if err == nil {
-			break
-		}
-
-		time.Sleep(1 * time.Second)
-	}
-
-	return resp, err
-}
-
-func GetTxResponseAndCheckLogs(umeeClient client.Client, txHash string) (*sdk.TxResponse, error) {
-	fullResp, err := GetTxResponse(umeeClient, txHash)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(fullResp.Logs) == 0 {
-		return nil, fmt.Errorf("no logs in response")
-	}
-
-	return fullResp, nil
-}
